renamed PatternTypes to symbols
