+++ src/symbols.ts
