+++ src/types/FindSelected.ts
@@ -1,4 +1,4 @@
-import type * as PatternType from '../PatternType';
+import type * as PatternType from '../symbols';
 import type { Cast, IsPlainObject, UnionToIntersection } from './helpers';
 import type { NamedSelectPattern, AnonymousSelectPattern } from './Pattern';
